Day 2: System clock created

+++ oss.h
@@ -4,7 +4,7 @@
 struct PCB {
 int occupied; // either true or false
 pid_t pid; // process id of this child
-int sec; // time when it was forked
-int nano; // time when it was forked
+double sec; // time when it was forked
+double nano; // time when it was forked
 };
 struct PCB processTable[20];
